feat(unity): Add IL2CPP v2022.3.56f1 support and refactor key finding
Adds support for Unity IL2CPP version 2022.3.56f1 alongside the existing 2022.3.22f1 support, selectable via Cargo features.

Refactors the IL2CPP metadata key extraction mechanism to parse ARM64 instructions (BL, ADRP, ADD immediate) instead of relying on less robust hex pattern matching. This improves reliability across different IL2CPP builds.

Key changes include:
- Introduced `2022322f1` and `2022356f1` features (`2022356f1` default).
- Added generated bindings for `2022.3.56f1` and updated `2022.3.22f1`.
- Added build scripts for header processing and binding generation.
- Implemented parsing for ARM64 BL, ADRP, and ADD immediate instructions.
- Added ELF instruction chunk parsing (`Elf::instructions`).
- Refactored `Il2Cpp::load_from_vec` to accept pre-parsed `Elf`.
- Updated metadata loading to use feature-flagged version constant.
- Modified key finding logic in `Il2Cpp` and `unity_loader`.

+++ src/unity/il2cpp_type.rs
@@ -208,7 +208,7 @@ impl<'a> ReadOnly<&'a Il2CppType> {
             },
             _ => {
                 // Use a predefined mapping for basic types. If not found, label as unknown.
-                if let Some(ty_str) = TYPE_MAP.get(&ty) {
+                if let Some(ty_str) = TYPE_MAP.get(&(ty as i32)) {
                     Ok(ComplexType::Simple {
                         module: None,
                         namespace: None,
